+++ icechunk/src/config.rs
@@ -466,7 +466,7 @@ pub struct RepositoryConfig {
     pub repo_update_retries: Option<RepoUpdateRetryConfig>,
 
     #[serde(default)]
-    pub num_updates_per_repo_info_file: Option<usize>,
+    pub num_updates_per_repo_info_file: Option<u16>,
 }
 
 static DEFAULT_COMPRESSION: OnceLock<CompressionConfig> = OnceLock::new();
@@ -475,8 +475,7 @@ static DEFAULT_MANIFEST_CONFIG: OnceLock<ManifestConfig> = OnceLock::new();
 static DEFAULT_REPO_UPDATE_RETRY_CONFIG: OnceLock<RepoUpdateRetryConfig> =
     OnceLock::new();
 pub const DEFAULT_MAX_CONCURRENT_REQUESTS: u16 = 256;
-// usize?
-pub const DEFAULT_NUM_UPDATES_PER_REPO_INFO_FILE: usize = 100;
+pub const DEFAULT_NUM_UPDATES_PER_REPO_INFO_FILE: u16 = 100;
 
 impl RepositoryConfig {
     pub fn inline_chunk_threshold_bytes(&self) -> u16 {
@@ -516,7 +515,7 @@ impl RepositoryConfig {
         })
     }
 
-    pub fn num_updates_per_repo_info_file(&self) -> usize {
+    pub fn num_updates_per_repo_info_file(&self) -> u16 {
         self.num_updates_per_repo_info_file
             .unwrap_or(DEFAULT_NUM_UPDATES_PER_REPO_INFO_FILE)
     }
